requirements.txt
numpy
scipy
pillow
pandas
matplotlib
scikit-learn
six
natsort